requirements3.5.txt
altgraph==0.17
asgiref==3.4.1
Django==2.2.24
django-adminlte3==0.1.6
django-filter==2.4.0
django-stdimage==5.3.0
Pillow==7.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
